AOC 2023 Day 03 - Part 02
Given an engine schematic of numbers, symbols, and periods, find the
part numbers, which are numbers adjacent to symbols, even diagonally
adjacent. A '*' symbol with two adjacent part numbers is a gear. Its
gear ratio is the product of those two part numbers. Sum all the gear
ratios in the engine schematic.